media-db: Port to GFile and add support for URIs in the database
Add support for URIs in the database, in a backwards-compatible manner. URIs
are stored in the files table with base_path_id set to 0, and no corresponding
entry in the paths table. Normal paths or local URIs are stored split between
the files and paths tables as before.

This adds a dependency on GIO, which was already implicit due to depending on
GLib 2.30.